expose StepT[F[_], A] and Step[A]
The type alias for the `StepT` monad transformer helps with using
`liftM[G: MonadTrans]` to lift a `Future[A]` directly to `Step[A]`.

The type alias for the `Step` monad helps with using `point[M: Monad]`
to lift an `A` directly to `Step[A]`.

See the ExampleController test for example usage.